Don't include the database id when exporting patients
The internal database id should not be used outside of the Transmart server, the trial id (inTrialId) should be used for that. The database id
is not useful anyway since it is not exported in other places so it cannot be correlated.

The accompanying update in RInterface is in the patient_sets branch.